Getting library to build and push correctly
