+++ extractors-opendronemap/Dockerfile
@@ -19,7 +19,7 @@ RUN pip install --upgrade pip && \
     pip install --ignore-installed enum34 terrautils
 
 # Copy and install our utilities
-RUN pip install pipelineutils 
+RUN pip install -U pipelineutils 
 
 # Copy our source files
 COPY entrypoint.sh extractor_info.json *.py /home/extractor/
